Add new view model for Qr
