+++ WalletWasabi.Fluent/Views/Dialogs/ShowQrCameraDialogView.axaml.cs
@@ -1,0 +1,22 @@
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace WalletWasabi.Fluent.Views.Dialogs
+{
+	public class ShowQrCameraDialogView : UserControl
+	{
+		// NASTY HACK! DO NOT USE IN PRODUCTION. kiminuo
+		public static volatile Image? QrImage;
+
+		public ShowQrCameraDialogView()
+		{
+			InitializeComponent();
+			QrImage = this.FindControl<Image>("qrImage");
+		}
+
+		private void InitializeComponent()
+		{
+			AvaloniaXamlLoader.Load(this);
+		}
+	}
+}
