+++ WalletWasabi.Fluent/Views/Wallets/Send/SendView.axaml
@@ -49,12 +49,6 @@
           </TextBox.InnerRightContent>
         </TextBox>
       </DockPanel>
-      <!-- QR Camera -->
-      <DockPanel  DockPanel.Dock="Top" Margin="0,0,0,6" IsVisible="{Binding IsQrPanelVisible}">
-        <Label DockPanel.Dock="Left" Content="_QR Camera:" Target="qrcamera" />
-        <Label DockPanel.Dock="Left" IsVisible="{Binding IsCameraLoadingAnimationVisible}" Content="_Loading camera..." Opacity="0.6"></Label>
-        <Image Source="{Binding QrImage}" Width="275" Height="275" Name="qrImage" />
-      </DockPanel>
       <!-- Amount -->
       <DockPanel DockPanel.Dock="Top" Margin="0,0,0,10">
         <Label DockPanel.Dock="Left" Content="_Amount:" Target="amountTb" />
