+++ WalletWasabi.Fluent/Views/Wallets/Send/SendView.axaml.cs
@@ -5,13 +5,9 @@ namespace WalletWasabi.Fluent.Views.Wallets.Send
 {
 	public class SendView : UserControl
 	{
-		// NASTY HACK! DO NOT USE IN PRODUCTION. kiminuo
-		public static volatile Image? QrImage;
-
 		public SendView()
 		{
 			InitializeComponent();
-			QrImage = this.FindControl<Image>("qrImage");
 		}
 
 		private void InitializeComponent()
